+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+- Fix configuration for check-mysql-query-result-count.rb script (@athal7)
 
 ## [1.2.0] - 2017-03-23
 - Add check-mysql-query-result-count.rb script (@athal7)